+++ src/main/resources/templates/member/selectAddress.html
@@ -1,0 +1,106 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"
+      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+      layout:decorate="~{layouts/layout2}">
+
+<head>
+  <meta name="_csrf" th:content="${_csrf.token}"/>
+  <meta name="_csrf_header" th:content="${_csrf.headerName}"/>
+</head>
+
+<!-- 사용자 스크립트 추가 -->
+<th:block layout:fragment="script">
+  <script th:inline="javascript">
+    $(document).ready(function () {
+      var errorMessage = [[${errorMessage}]];
+      if (errorMessage != null) {
+        alert(errorMessage);
+      }
+    });
+  </script>
+
+  <script>
+    function sendSelectedId() {
+      const selectedId = document.querySelector('input[name="addressIdGroup"]:checked').value;
+
+      window.opener.postMessage({ type: 'selectedId', selectedId }, '*');
+
+      window.close();
+    }
+  </script>
+
+
+</th:block>
+
+<!-- 사용자 CSS 추가 -->
+<th:block layout:fragment="css">
+  <style>
+    .content-mg {
+      margin-left: 25%;
+      margin-right: 25%;
+      margin-top: 2%;
+      margin-bottom: 100px;
+    }
+
+    .order-box {
+      width: 700px;
+    }
+
+    .order-line {
+      border-bottom: 1px solid #efefef;
+    }
+
+    .fieldError {
+      color: #bd2130;
+      font-size: 0.875rem;
+      line-height: 1.25rem;
+      width: 100%;
+      padding-top: 0.5rem;
+      padding-left: 0.5rem;
+    }
+  </style>
+</th:block>
+
+<div layout:fragment="content" class="content-mg">
+
+  <div class="flex w-full justify-center">
+    <div class="flex flex-col" style="width: 650px;">
+            <span class="text-3xl font-bold my-4 ml-3">
+                배송지 변경
+                <a type="button" class="btn btn-accent float-right ml-2"
+                   th:href="@{/members/mypage/address/new}"
+                   th:text="|신규 배송지|"></a>
+              <button type="button" class="btn btn-primary float-right mb-3 ml-1"
+                      onclick="sendSelectedId()">배송지 변경</button>
+            </span>
+      <div th:each="address : ${addressList}" class="flex mb-3">
+        <div class="order-box card bg-base-100 shadow-xl">
+          <div class="card-body">
+            <div class="join-form flex justify-center flex-col">
+              <dl>
+                <dt class="detail-title mb-2">
+                  <input type="radio" class="form-check-input" th:value="${address.addressId}" name="addressIdGroup">
+                  <span class="font-bold" th:text="${address.recipientName}"></span>
+                  <span class="font-bold" th:text="${address.shippingName}"></span>
+                  <span th:if="${address.is_default == true}">기본 배송지</span>
+                </dt>
+                <dd class="detail-content mb-3">
+                  <div th:text="${address.addr + ' ' + address.addr_detail}"></div>
+                  <div th:text="${address.recipientContact}"></div>
+                </dd>
+                <a type="button" class="btn btn-accent float-right ml-2"
+                   th:href="@{|/members/mypage/address/modify/${address.addressId}|}"
+                   th:text="|수정|"></a>
+                <a type="button" class="btn btn-error float-right"
+                   th:href="@{|/members/mypage/address/delete/${address.addressId}|}"
+                   th:text="|삭제|"></a>
+              </dl>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
+
+</html>
