+++ src/main/resources/templates/order/orderDtl.html
@@ -25,6 +25,7 @@
     <script>
         var urlParams = new URLSearchParams(window.location.search);
         var msg = urlParams.get('msg');
+        let selectedId = null;
 
         if (msg) {
             alert(decodeURIComponent(msg));
@@ -64,16 +65,15 @@
                 },
                 dataType: "json",
                 cache: false,
-                success: function (result, status) {
+                success: function () {
                     payment();
                 },
-                error: function (jqXHR, status, error) {
+                error: function (jqXHR) {
 
                     if (jqXHR.status == '401') {
                         alert('로그인 후 이용해주세요');
                         location.href = '/members/login';
                     } else {
-                        // alert(jqXHR.responseText);
                         var errors = jqXHR.responseJSON;
                         displayErrors(errors);
                     }
@@ -81,8 +81,46 @@
             });
         }
 
+        function selectAddress() {
+            const popOption = "width = 650px, height=550px, top=300px, left=300px, scrollbars=yes";
+            const openUrl = '/members/mypage/address/select';
+            const childWindow = window.open(openUrl, 'select', popOption);
+
+            window.addEventListener('message', function(event) {
+                if (event.data.type === 'selectedId') {
+                    selectedId = event.data.selectedId;
+                    receiveSelectedId(selectedId);
+
+                    childWindow.close();
+                }
+            });
+        }
+
+        function receiveSelectedId(selectedId) {
+            const apiEndpoint = baseUrl + `/members/mypage/address/select/${selectedId}`
+            fetch(apiEndpoint)
+                .then(response => response.json())
+                .then(data => {
+                    const shippingName = document.getElementById('shippingName');
+                    const recipientName = document.getElementById('recipientName');
+                    const recipientContact = document.getElementById('recipientContact');
+                    const zipcode = document.getElementById('zipcode');
+                    const addr = document.getElementById('addr');
+                    const addr_detail = document.getElementById('addr_detail');
+                    shippingName.value = data.shippingName;
+                    recipientName.value = data.recipientName;
+                    recipientContact.value = data.recipientContact;
+                    zipcode.value = data.zipcode;
+                    addr.value = data.addr;
+                    addr_detail.value = data.addr_detail;
+                })
+                .catch(error => {
+                    alert(error);
+                });
+        }
+
         function displayErrors(errors) {
-            $(".fieldError").hide().text(""); // Clear previous error messages
+            $(".fieldError").hide().text("");
 
             for (var fieldName in errors) {
                 var errorElement = $("#" + fieldName + "Error");
@@ -92,11 +130,11 @@
                     if (errors[fieldName]) {
                         errorElement.text(errors[fieldName]);
                         errorElement.show();
-                        inputElement.addClass("is-invalid"); // Add error styling
+                        inputElement.addClass("is-invalid");
                     } else {
                         errorElement.hide();
-                        inputElement.removeClass("is-invalid"); // Remove error styling
-                        inputElement.val(""); // Clear input value
+                        inputElement.removeClass("is-invalid");
+                        inputElement.val("");
                     }
                 }
             }
@@ -130,54 +168,6 @@
             };
 
             tossPayments.requestPayment(method, paymentData);
-        }
-
-        function execDaumPostcode() {
-            new daum.Postcode({
-                oncomplete: function (data) {
-                    // 팝업에서 검색결과 항목을 클릭했을때 실행할 코드를 작성하는 부분.
-
-                    // 각 주소의 노출 규칙에 따라 주소를 조합한다.
-                    // 내려오는 변수가 값이 없는 경우엔 공백('')값을 가지므로, 이를 참고하여 분기 한다.
-                    var addr = ''; // 주소 변수
-                    var extraAddr = ''; // 참고항목 변수
-
-                    //사용자가 선택한 주소 타입에 따라 해당 주소 값을 가져온다.
-                    if (data.userSelectedType === 'R') { // 사용자가 도로명 주소를 선택했을 경우
-                        addr = data.roadAddress;
-                    } else { // 사용자가 지번 주소를 선택했을 경우(J)
-                        addr = data.jibunAddress;
-                    }
-
-                    // 사용자가 선택한 주소가 도로명 타입일때 참고항목을 조합한다.
-                    if (data.userSelectedType === 'R') {
-                        // 법정동명이 있을 경우 추가한다. (법정리는 제외)
-                        // 법정동의 경우 마지막 문자가 "동/로/가"로 끝난다.
-                        if (data.bname !== '' && /[동|로|가]$/g.test(data.bname)) {
-                            extraAddr += data.bname;
-                        }
-                        // 건물명이 있고, 공동주택일 경우 추가한다.
-                        if (data.buildingName !== '' && data.apartment === 'Y') {
-                            extraAddr += (extraAddr !== '' ? ', ' + data.buildingName : data.buildingName);
-                        }
-                        // 표시할 참고항목이 있을 경우, 괄호까지 추가한 최종 문자열을 만든다.
-                        if (extraAddr !== '') {
-                            extraAddr = ' (' + extraAddr + ')';
-                        }
-                        // 조합된 참고항목을 해당 필드에 넣는다.
-                        document.getElementById("addr_detail").value = extraAddr;
-
-                    } else {
-                        document.getElementById("addr_detail").value = '';
-                    }
-
-                    // 우편번호와 주소 정보를 해당 필드에 넣는다.
-                    document.getElementById('zipcode').value = data.zonecode;
-                    document.getElementById("addr").value = addr;
-                    // 커서를 상세주소 필드로 이동한다.
-                    document.getElementById("addr_detail").focus();
-                }
-            }).open();
         }
     </script>
 </th:block>
@@ -217,6 +207,9 @@
             width:100%;
             padding-top: 0.5rem;
             padding-left: 0.5rem;
+        }
+        #inlineh3 {
+            display: inline-block;
         }
     </style>
 </th:block>
@@ -280,7 +273,11 @@
                                 </dd>
                             </dl>
                         </div>
-                        <h3 class="card-title text-2xl">받으시는 분</h3>
+                        <div class="d-inline-flex">
+                            <h3 class="card-title text-2xl" id="inlineh3">받으시는 분</h3>
+                                    <button type="button" class="btn btn-primary float-right mb-3 ml-1"
+                                            onclick="selectAddress()">배송지 변경</button>
+                        </div>
                         <div class="join-form flex justify-center flex-col">
                             <dl class="mb-3">
                                 <dt class="detail-title">
@@ -319,9 +316,6 @@
                                 <dd class="detail-content">
                                     <input type="text" id="zipcode" class="input input-bordered"
                                            placeholder="우편번호" readonly>
-                                    <button type="button" class="btn btn-primary mb-3 ml-1"
-                                            onclick="execDaumPostcode()">주소검색
-                                    </button>
                                     <div id="zipcodeError" class="fieldError"></div>
                                 </dd>
                             </dl>
